# storageapi/restapi/op/run_despenser_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package op

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"
	"github.com/go-openapi/runtime/middleware"
)

// RunDespenserNoContentCode is the HTTP code returned for type RunDespenserNoContent
const RunDespenserNoContentCode int = 204

/*RunDespenserNoContent OK

swagger:response runDespenserNoContent
*/
type RunDespenserNoContent struct {
}

// NewRunDespenserNoContent creates RunDespenserNoContent with default headers values
func NewRunDespenserNoContent() *RunDespenserNoContent {

	return &RunDespenserNoContent{}
}

// WriteResponse to the client
func (o *RunDespenserNoContent) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(204)
}

func (o *RunDespenserNoContent) RunDespenserResponder() {}

// RunDespenserNotFoundCode is the HTTP code returned for type RunDespenserNotFound
const RunDespenserNotFoundCode int = 404

/*RunDespenserNotFound not found

swagger:response runDespenserNotFound
*/
type RunDespenserNotFound struct {

	/*
	  In: Body
	*/
	Payload string `json:"body,omitempty"`
}

// NewRunDespenserNotFound creates RunDespenserNotFound with default headers values
func NewRunDespenserNotFound() *RunDespenserNotFound {

	return &RunDespenserNotFound{}
}

// WithPayload adds the payload to the run despenser not found response
func (o *RunDespenserNotFound) WithPayload(payload string) *RunDespenserNotFound {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the run despenser not found response
func (o *RunDespenserNotFound) SetPayload(payload string) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *RunDespenserNotFound) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(404)
	payload := o.Payload
	if err := producer.Produce(rw, payload); err != nil {
		panic(err) // let the recovery middleware deal with this
	}
}

func (o *RunDespenserNotFound) RunDespenserResponder() {}

// RunDespenserInternalServerErrorCode is the HTTP code returned for type RunDespenserInternalServerError
const RunDespenserInternalServerErrorCode int = 500

/*RunDespenserInternalServerError internal error

swagger:response runDespenserInternalServerError
*/
type RunDespenserInternalServerError struct {
}

// NewRunDespenserInternalServerError creates RunDespenserInternalServerError with default headers values
func NewRunDespenserInternalServerError() *RunDespenserInternalServerError {

	return &RunDespenserInternalServerError{}
}

// WriteResponse to the client
func (o *RunDespenserInternalServerError) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(500)
}

func (o *RunDespenserInternalServerError) RunDespenserResponder() {}

type RunDespenserNotImplementedResponder struct {
	middleware.Responder
}

func (*RunDespenserNotImplementedResponder) RunDespenserResponder() {}

func RunDespenserNotImplemented() RunDespenserResponder {
	return &RunDespenserNotImplementedResponder{
		middleware.NotImplemented(
			"operation authentication.RunDespenser has not yet been implemented",
		),
	}
}

type RunDespenserResponder interface {
	middleware.Responder
	RunDespenserResponder()
}