# storageapi/restapi/op/volume_despenser_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package op

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"
	"github.com/go-openapi/runtime/middleware"
)

// VolumeDespenserOKCode is the HTTP code returned for type VolumeDespenserOK
const VolumeDespenserOKCode int = 200

/*VolumeDespenserOK OK

swagger:response volumeDespenserOK
*/
type VolumeDespenserOK struct {

	/*
	  In: Body
	*/
	Payload *VolumeDespenserOKBody `json:"body,omitempty"`
}

// NewVolumeDespenserOK creates VolumeDespenserOK with default headers values
func NewVolumeDespenserOK() *VolumeDespenserOK {

	return &VolumeDespenserOK{}
}

// WithPayload adds the payload to the volume despenser o k response
func (o *VolumeDespenserOK) WithPayload(payload *VolumeDespenserOKBody) *VolumeDespenserOK {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the volume despenser o k response
func (o *VolumeDespenserOK) SetPayload(payload *VolumeDespenserOKBody) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *VolumeDespenserOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(200)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}

func (o *VolumeDespenserOK) VolumeDespenserResponder() {}

// VolumeDespenserNoContentCode is the HTTP code returned for type VolumeDespenserNoContent
const VolumeDespenserNoContentCode int = 204

/*VolumeDespenserNoContent OK

swagger:response volumeDespenserNoContent
*/
type VolumeDespenserNoContent struct {
}

// NewVolumeDespenserNoContent creates VolumeDespenserNoContent with default headers values
func NewVolumeDespenserNoContent() *VolumeDespenserNoContent {

	return &VolumeDespenserNoContent{}
}

// WriteResponse to the client
func (o *VolumeDespenserNoContent) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(204)
}

func (o *VolumeDespenserNoContent) VolumeDespenserResponder() {}

// VolumeDespenserNotFoundCode is the HTTP code returned for type VolumeDespenserNotFound
const VolumeDespenserNotFoundCode int = 404

/*VolumeDespenserNotFound not found

swagger:response volumeDespenserNotFound
*/
type VolumeDespenserNotFound struct {

	/*
	  In: Body
	*/
	Payload string `json:"body,omitempty"`
}

// NewVolumeDespenserNotFound creates VolumeDespenserNotFound with default headers values
func NewVolumeDespenserNotFound() *VolumeDespenserNotFound {

	return &VolumeDespenserNotFound{}
}

// WithPayload adds the payload to the volume despenser not found response
func (o *VolumeDespenserNotFound) WithPayload(payload string) *VolumeDespenserNotFound {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the volume despenser not found response
func (o *VolumeDespenserNotFound) SetPayload(payload string) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *VolumeDespenserNotFound) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(404)
	payload := o.Payload
	if err := producer.Produce(rw, payload); err != nil {
		panic(err) // let the recovery middleware deal with this
	}
}

func (o *VolumeDespenserNotFound) VolumeDespenserResponder() {}

// VolumeDespenserInternalServerErrorCode is the HTTP code returned for type VolumeDespenserInternalServerError
const VolumeDespenserInternalServerErrorCode int = 500

/*VolumeDespenserInternalServerError internal error

swagger:response volumeDespenserInternalServerError
*/
type VolumeDespenserInternalServerError struct {
}

// NewVolumeDespenserInternalServerError creates VolumeDespenserInternalServerError with default headers values
func NewVolumeDespenserInternalServerError() *VolumeDespenserInternalServerError {

	return &VolumeDespenserInternalServerError{}
}

// WriteResponse to the client
func (o *VolumeDespenserInternalServerError) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(500)
}

func (o *VolumeDespenserInternalServerError) VolumeDespenserResponder() {}

type VolumeDespenserNotImplementedResponder struct {
	middleware.Responder
}

func (*VolumeDespenserNotImplementedResponder) VolumeDespenserResponder() {}

func VolumeDespenserNotImplemented() VolumeDespenserResponder {
	return &VolumeDespenserNotImplementedResponder{
		middleware.NotImplemented(
			"operation authentication.VolumeDespenser has not yet been implemented",
		),
	}
}

type VolumeDespenserResponder interface {
	middleware.Responder
	VolumeDespenserResponder()
}