# storageapi/swagger.yml
swagger: "2.0"
info:
  title: Storage
  version: 1.8.0
basePath: /
schemes:
  - http
securityDefinitions:
   pinCode:
     type: apiKey
     in: header
     name: Pin
definitions:
...
  Hash:
    type: string
    minLength: 1
...
consumes:
  - application/json
produces:
  - application/json          
paths:
...
  /run-despenser:
    post:
      operationId: "runDespenser"
      parameters:
        - name: args
          required: true
          in: body
          schema:
            title: "ArgRunDespenser"
            type: object
            required:
              - hash
              - volume
            properties:
              hash:
                $ref: '#/definitions/Hash'
              volume:
                type: integer
      responses:
        204:
          description: OK
        404:
          description: not found
          schema:
            type: string
        500:
          description: internal error
  /volume-despenser:
    post:
      operationId: "VolumeDespenser"
      parameters:
        - name: args
          required: true
          in: body
          schema:
            title: "VolumeDespenser"
            type: object
            required:
              - hash
            properties:
              hash:
                $ref: '#/definitions/Hash'
      responses:
        200:
          description: OK
          schema:
            title: "ResponseVolumeDespenser"
            type: object
            required:
              - volume
            properties:
              volume: 
                type: integer
        204:
          description: OK
        404:
          description: not found
          schema:
            type: string
        500:
          description: internal error
...